% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonMisesM.R
\name{dvonMisesM}
\alias{dvonMisesM}
\title{Density function for the von Mises mixture distribution}
\usage{
dvonMisesM(x, mu, kappa, theta)
}
\arguments{
\item{x}{random variables.}

\item{mu}{mean direction of each component distribution.}

\item{kappa}{non-negative numeric value for the concentration parameter of each component distribution.}

\item{theta}{simplex vector for mixing ratios}
}
\description{
\code{dvonMisesM} Density function for the von Mises mixture distribution
}
\examples{
#rvonMisesM(pi, c(2.1, 3.4), c(1.0, 1.0), c(0.5, 0.5))
}
