% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{fit.EM}
\alias{fit.EM}
\title{Parameter estimation of von Mises Mixture model using EM algorithm}
\usage{
fit.EM(N_step, K, y)
}
\arguments{
\item{N_step}{XXXX.}

\item{K}{XXX.}

\item{y}{XXX.}
}
\value{
XXX.
}
\description{
\code{fit.EM} XXXX
}
\examples{
#fit.EM(10, K, y)
}
