% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonMisesM.R
\name{rvonMisesM}
\alias{rvonMisesM}
\title{Random number generation from von Mises mixture distribution}
\usage{
rvonMisesM(n, mu, kappa, theta)
}
\arguments{
\item{n}{number of observations.}

\item{mu}{mean direction of each component distribution.}

\item{kappa}{non-negative numeric value for the concentration parameter of each component distribution.}

\item{theta}{simplex vector for mixing ratios.}
}
\value{
Random number of length n.
}
\description{
\code{rvonMisesM} Generate random numbers from von Mises mixture distribution
}
\examples{
#rvonMisesM(10, c(2.1, 3.4), c(1.0, 1.0), c(0.5, 0.5))
}
